postgres-elasticsearch/requirements.txt
psycopg2-binary==2.9.9
elasticsearch==7.17.9
confluent-kafka==2.3.0
python-dotenv==1.0.0